feat/update-packages: update versions and package files
