+++ components/package.json
@@ -17,21 +17,25 @@
     "lint": "eslint . src/**/*.ts",
     "lint:fix": "eslint . src/**/*.ts --fix"
   },
+  "engines": {
+    "node": "16.19.0",
+    "npm": "8.19.3"
+  },
   "dependencies": {
     "@aotw/core": "file:../core",
-    "lit": "^2.2.2"
+    "lit": "2.5.0"
   },
   "devDependencies": {
     "@aotw/eslint-config": "file:../eslint-config",
-    "@types/node": "^17.0.23",
-    "@typescript-eslint/eslint-plugin": "^5.33.0",
-    "@typescript-eslint/parser": "^5.33.0",
-    "eslint": "^8.22.0",
-    "eslint-plugin-lit": "^1.6.1",
-    "rimraf": "^3.0.2",
-    "rollup-plugin-postcss-lit": "^2.0.0",
-    "sass": "^1.50.0",
-    "typescript": "^4.5.4",
-    "vite": "^2.9.0"
+    "@types/node": "18.11.9",
+    "@typescript-eslint/eslint-plugin": "5.47.1",
+    "@typescript-eslint/parser": "5.47.1",
+    "eslint": "8.30.0",
+    "eslint-plugin-lit": "1.7.2",
+    "rimraf": "3.0.2",
+    "rollup-plugin-postcss-lit": "2.0.0",
+    "sass": "1.57.1",
+    "typescript": "4.9.4",
+    "vite": "4.0.3"
   }
 }
