--- components/types/button/button.model.d.ts
@@ -1,4 +1,0 @@
-export declare enum ButtonVariant {
-    Primary = "primary",
-    Secondary = "secondary"
-}
