--- components/types/button/index.d.ts
@@ -1,2 +1,0 @@
-export * from './button';
-export * from './button.model';
