+++ components/types/container/container.d.ts
@@ -1,0 +1,6 @@
+import { LitElement, TemplateResult } from 'lit';
+export declare class AotwContainer extends LitElement {
+    elevated: boolean;
+    static styles: import("lit").CSSResult[];
+    protected render(): TemplateResult;
+}
