+++ components/types/container/index.d.ts
@@ -1,0 +1,1 @@
+export * from './container';
