+++ components/types/field/field/field.d.ts
@@ -1,8 +1,5 @@
 import { LitElement, TemplateResult } from 'lit';
 export declare class AotwField extends LitElement {
-    private _disabled;
-    set disabled(disabled: boolean);
-    get disabled(): boolean;
     private _elements;
     static styles: import("lit").CSSResult;
     protected firstUpdated(): void;
