--- components/types/icon/icons/icon-arrow-down.d.ts
@@ -1,4 +1,0 @@
-import { LitElement } from 'lit';
-export declare class IconArrowDownElement extends LitElement {
-    protected render(): import("lit-html").TemplateResult<2>;
-}
