--- components/types/icon/icons/icon-arrow-up.d.ts
@@ -1,4 +1,0 @@
-import { LitElement } from 'lit';
-export declare class IconArrowUpElement extends LitElement {
-    protected render(): import("lit-html").TemplateResult<2>;
-}
