--- components/types/icon/icons/icon-check.d.ts
@@ -1,4 +1,0 @@
-import { LitElement } from 'lit';
-export declare class IconCheckElement extends LitElement {
-    protected render(): import("lit-html").TemplateResult<2>;
-}
