--- components/types/icon/icons/icon-close.d.ts
@@ -1,4 +1,0 @@
-import { LitElement } from 'lit';
-export declare class IconCloseElement extends LitElement {
-    protected render(): import("lit-html").TemplateResult<2>;
-}
