--- components/types/icon/icons/index.d.ts
@@ -1,7 +1,0 @@
-export * from './icon-arrow-down';
-export * from './icon-arrow-up';
-export * from './icon-check';
-export * from './icon-close';
-export * from './icon-dot';
-export * from './icon-logo';
-export * from './icon-shine';
