+++ components/types/index.d.ts
@@ -1,13 +1,12 @@
-export * from './button';
 export * from './chip';
 export * from './close-button';
+export * from './container';
 export * from './dialog';
 export * from './dynamic-text';
 export * from './icon';
 export * from './field';
-export * from './list';
+export * from './list-item';
 export * from './overlay';
 export * from './tab';
 export * from './table';
 export * from './tag';
-export * from './toggle-switch';
