+++ components/types/list-item/index.d.ts
@@ -1,2 +1,1 @@
-export * from './list';
 export * from './list-item';
