--- components/types/list/list.d.ts
@@ -1,10 +1,0 @@
-import { LitElement, TemplateResult } from 'lit';
-export declare class ListElement extends LitElement {
-    static styles: import("lit").CSSResult;
-    protected render(): TemplateResult;
-}
-declare global {
-    interface HTMLElementTagNameMap {
-        AOTW_LIST: ListElement;
-    }
-}
