--- components/types/toggle-switch/index.d.ts
@@ -1,1 +1,0 @@
-export * from './toggle-switch';
