+++ components/typings/scss.d.ts
@@ -1,0 +1,5 @@
+declare module '*.scss?inline' {
+  import { CSSResult } from 'lit';
+  const styles: CSSResult;
+  export default styles;
+}
