+++ core/package.json
@@ -11,10 +11,15 @@
   "scripts": {
     "build:tokens": "style-dictionary build --config ./config.json",
     "build:scss": "sass src/index.scss:dist/style.css",
-    "build": "npm run build:tokens && npm run build:scss"
+    "build": "rimraf dist && npm run build:tokens && npm run build:scss"
+  },
+  "engines": {
+    "node": "16.19.0",
+    "npm": "8.19.3"
   },
   "devDependencies": {
-    "sass": "^1.50.0",
-    "style-dictionary": "^3.7.1"
+    "rimraf": "3.0.2",
+    "sass": "1.57.1",
+    "style-dictionary": "3.7.1"
   }
 }
