+++ core/src/tokens/variables.css
@@ -1,6 +1,6 @@
 /**
  * Do not edit directly
- * Generated on Tue, 20 Dec 2022 14:10:20 GMT
+ * Generated on Wed, 28 Dec 2022 19:51:42 GMT
  */
 
 :root {
