+++ eslint-config/package.json
@@ -6,13 +6,17 @@
   "license": "ISC",
   "main": ".eslintrc",
   "scripts": {},
-  "devDependencies": {
-    "eslint-config-prettier": "^8.5.0"
+  "engines": {
+    "node": "16.19.0",
+    "npm": "8.19.3"
   },
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.41.0",
-    "@typescript-eslint/parser": "^5.41.0",
-    "eslint": "^8.26.0",
-    "typescript": "^4.8.4"
+    "@typescript-eslint/eslint-plugin": "5.47.1",
+    "@typescript-eslint/parser": "5.47.1",
+    "eslint": "8.30.0",
+    "typescript": "4.9.4"
+  },
+  "devDependencies": {
+    "eslint-config-prettier": "8.5.0"
   }
 }
