+++ storybook/.npmrc
@@ -1,0 +1,1 @@
+engine-strict=true
