+++ storybook/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@aotw/storybook",
   "description": "Archive of the World Library Storybook",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "author": "Jurrit van der Ploeg <info@jurritvanderploeg.nl> (http://www.jurritvanderploeg.nl/)",
   "license": "MIT",
   "private": true,
@@ -11,26 +11,30 @@
     "lint": "eslint . src/**/*.ts",
     "lint:fix": "eslint . src/**/*.ts --fix"
   },
+  "engines": {
+    "node": "16.19.0",
+    "npm": "8.19.3"
+  },
   "dependencies": {
     "@aotw/components": "file:../components",
     "@aotw/core": "file:../core",
-    "lit": "^2.2.2"
+    "lit": "2.5.0"
   },
   "devDependencies": {
     "@aotw/eslint-config": "file:../eslint-config",
-    "@storybook/addon-actions": "^6.4.22",
-    "@storybook/addon-essentials": "^6.4.22",
-    "@storybook/addon-links": "^6.4.22",
-    "@storybook/addons": "^6.5.12",
-    "@storybook/builder-vite": "^0.1.29",
-    "@storybook/theming": "^6.5.12",
-    "@storybook/web-components": "^6.4.22",
-    "@typescript-eslint/eslint-plugin": "^5.41.0",
-    "eslint-plugin-lit": "^1.6.1",
-    "rollup-plugin-postcss-lit": "^2.0.0",
-    "sass": "^1.50.1",
-    "typescript": "^4.6.3",
-    "vite": "^2.9.5"
+    "@storybook/addon-actions": "6.5.15",
+    "@storybook/addon-essentials": "6.5.15",
+    "@storybook/addon-links": "6.5.15",
+    "@storybook/addons": "6.5.15",
+    "@storybook/builder-vite": "0.2.6",
+    "@storybook/theming": "6.5.15",
+    "@storybook/web-components": "6.5.15",
+    "@typescript-eslint/eslint-plugin": "5.47.1",
+    "eslint-plugin-lit": "1.7.2",
+    "rollup-plugin-postcss-lit": "2.0.0",
+    "sass": "1.57.1",
+    "typescript": "4.9.4",
+    "vite": "4.0.3"
   },
   "overrides": {
     "react": "^17.0.0",
